--- a/mlx/io/load.cpp
+++ b/mlx/io/load.cpp
@@ -298,7 +298,44 @@ array load(std::shared_ptr<io::Reader> in_stream, StreamOrDevice s) {
 
 /** Load array from file in .npy format */
 array load(std::string file, StreamOrDevice s) {
-  return load(std::make_shared<io::FileReader>(std::move(file)), s);
+  return load(std::make_shared<io::ParallelFileReader>(std::move(file), 4), s);
 }
+
+namespace io {
+
+void ParallelFileReader::read(char* data, size_t n) {
+  while (n != 0) {
+    auto m = ::read(fd_, data, std::min(n, static_cast<size_t>(INT32_MAX)));
+    if (m <= 0) {
+      std::ostringstream msg;
+      msg << "[read] Unable to read " << n << " bytes from file.";
+      throw std::runtime_error(msg.str());
+    }
+    data += m;
+    n -= m;
+  }
+}
+
+void ParallelFileReader::read(char* data, size_t n, size_t offset) {
+  auto readfn = [fd = fd_](size_t offset, size_t size, char* buffer) -> size_t {
+    return pread(fd, buffer, size, offset);
+  };
+  std::vector<std::future<size_t>> futs;
+  while (n != 0) {
+    size_t m = std::min(batch_size_, n);
+    futs.emplace_back(thread_pool_.enqueue(readfn, offset, m, data));
+    data += m;
+    n -= m;
+    offset += m;
+  }
+  for (auto& f : futs) {
+    size_t m = f.get();
+    if (m <= 0) {
+      throw std::runtime_error("[read] Unable to read from file.");
+    }
+  }
+}
+
+} // namespace io
 
 } // namespace mlx::core