--- a/mlx/io/safetensors.cpp
+++ b/mlx/io/safetensors.cpp
@@ -147,7 +147,7 @@ SafetensorsLoad load_safetensors(
 }
 
 SafetensorsLoad load_safetensors(const std::string& file, StreamOrDevice s) {
-  return load_safetensors(std::make_shared<io::FileReader>(file), s);
+  return load_safetensors(std::make_shared<io::ParallelFileReader>(file, 4), s);
 }
 
 void save_safetensors(